% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccounts.R
\name{getAccounts}
\alias{getAccounts}
\title{tcsinvest}
\usage{
getAccounts(token = "", sandbox = TRUE)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
