% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBalance.R
\name{getBalance}
\alias{getBalance}
\title{tcsinvest}
\usage{
getBalance(token = "", sandbox = TRUE)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
