% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistoricalData.R
\name{getHistoricalData}
\alias{getHistoricalData}
\title{Download Historical Bar Data}
\usage{
getHistoricalData(
  token = "",
  live = FALSE,
  figi = "",
  from = Sys.Date() - 2,
  to = Sys.Date(),
  interval = "hour",
  verbose = FALSE
)
}
\arguments{
\item{token}{token from Tinkoff account}

\item{live}{live trading - TRUE or sandbox (paper) trading - FALSE (default)}

\item{figi}{internal tinkoff code for instrument}

\item{from}{from what date download history}

\item{to}{to what date download history}

\item{interval}{timeframe of bars}

\item{verbose}{display status of retrieval (default FALSE)}
}
\description{
Makes a request to the server, and returns an data.table object with information about prices of instrument if successful.
}
\details{
As described by the official Tinkoff Investments documentation
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
token = 'your_sandbox_token_from_tcs_account'
getHistoricalData(token,figi = 'BBG005HLTYH9')
}
\seealso{
\code{\link{getOrderBook}}
}
\author{
Vyacheslav Arbuzov
}
