% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrders.R
\name{getOrders}
\alias{getOrders}
\title{Download all live orders in account}
\usage{
getOrders(token = "", sandbox = TRUE)
}
\arguments{
\item{token}{token from Tinkoff account}

\item{sandbox}{paper (TRUE) or live (FALSE) trading}
}
\description{
Makes a request to the server, and returns a data.table object with information about all live orders related with your account if successful.
}
\details{
As described by the official Tinkoff Investments documentation
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
token = 'your_token_from_tcs_account'
getOrders(token)
}
\seealso{
\code{\link{getTrades}} \code{\link{cancelOrder}}
}
\author{
Vyacheslav Arbuzov
}
