% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPortfolio.R
\name{getPortfolio}
\alias{getPortfolio}
\title{tcsinvest}
\usage{
getPortfolio(token = "", sandbox = TRUE)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
