% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuotes.R
\name{getQuotes}
\alias{getQuotes}
\title{tcsinvest}
\usage{
getQuotes(token = "", figi = "", sandbox = TRUE)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
