% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStocks.R
\name{getStocks}
\alias{getStocks}
\title{tcsinvest}
\usage{
getStocks(token = "", sandbox = TRUE)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
