% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbolInfo.R
\name{getSymbolInfo}
\alias{getSymbolInfo}
\title{Information for intrument by figi}
\usage{
getSymbolInfo(token = "", figi = "", sandbox = TRUE)
}
\arguments{
\item{token}{token from Tinkoff account}

\item{figi}{internal tinkoff code for instrument}

\item{sandbox}{paper (TRUE) or live (FALSE) trading}
}
\description{
Makes a request to the server, and returns an data.table object with information about instrument if successful.
}
\details{
As described by the official Tinkoff Investments documentation
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
token = 'your_token_from_tcs_account'
getSymbolInfo(token,figi = 'BBG005HLTYH9')
}
\seealso{
\code{\link{getTickerInfo}}
}
\author{
Vyacheslav Arbuzov
}
