% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrades.R
\name{getTrades}
\alias{getTrades}
\title{tcsinvest}
\usage{
getTrades(token = "", from = Sys.Date() - 2, to = Sys.Date(), sandbox = TRUE)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
