% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limitOrder.R
\name{limitOrder}
\alias{limitOrder}
\title{tcsinvest}
\usage{
limitOrder(
  token = "",
  figi = "",
  direction = NULL,
  lots = NULL,
  price = NULL,
  sandbox = TRUE
)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
