% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketOrder.R
\name{marketOrder}
\alias{marketOrder}
\title{tcsinvest}
\usage{
marketOrder(
  token = "",
  figi = "",
  direction = NULL,
  lots = NULL,
  sandbox = TRUE
)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
