% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandboxDeleteAccount.R
\name{sandboxDeleteAccount}
\alias{sandboxDeleteAccount}
\title{tcsinvest}
\usage{
sandboxDeleteAccount(token = "")
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
