% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandboxDeletePositions.R
\name{sandboxDeletePositions}
\alias{sandboxDeletePositions}
\title{Working with sandbox. Delete positions}
\usage{
sandboxDeletePositions(token = "")
}
\arguments{
\item{token}{token from Tinkoff account (only sandbox token)}
}
\description{
Delete sandbox positions (all positions: stocks, etfs,bonds and money)
}
\details{
If you need to create few money position, use this function for each position
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
token = 'your_token_from_tcs_account'
sandboxDeletePositions(token)
}
\seealso{
\code{\link{sandboxPositions}}
}
\author{
Vyacheslav Arbuzov
}
