% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandboxRegister.R
\name{sandboxRegister}
\alias{sandboxRegister}
\title{tcsinvest}
\usage{
sandboxRegister(token = "")
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
