% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamClient.R
\name{streamClient}
\alias{streamClient}
\title{Create stream client}
\usage{
streamClient(token = "")
}
\arguments{
\item{token}{token from Tinkoff account}
}
\description{
this function creates client for streaming
}
\details{
As described by the official Tinkoff Investments documentation. If you want live trading, use sandbox=FALSE with live token
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\seealso{
\code{\link{streamStatus}} \code{\link{streamClose}} \code{\link{streamSubscribe}}
}
\author{
Vyacheslav Arbuzov
}
