% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamStatus.R
\name{streamStatus}
\alias{streamStatus}
\title{Check stream status}
\usage{
streamStatus(client = NULL)
}
\arguments{
\item{client}{object from streamClient result, class of object ws}
}
\description{
this function return current status of connection
}
\details{
As described by the official Tinkoff Investments documentation. If you want live trading, use sandbox=FALSE with live token
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\seealso{
\code{\link{streamClose}}  \code{\link{streamClient}} \code{\link{streamSubscribe}}
}
\author{
Vyacheslav Arbuzov
}
