% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamSubscribe.R
\name{streamSubscribe}
\alias{streamSubscribe}
\title{Subscribe/Unsubscribe to stream}
\usage{
streamSubscribe(
  client = NULL,
  subscribe = TRUE,
  type = "candle",
  figi = "BBG004730N88",
  FUN = "print",
  interval = "1min",
  depth = 10
)
}
\arguments{
\item{client}{object from streamClient result, class of object ws}

\item{subscribe}{subscribe - TRUE (default) or unsubscribe - FALSE}

\item{type}{type of events for subscription: "candle","orderbook" or "instrument_info"}

\item{figi}{internal tinkoff code for instrument}

\item{FUN}{function for handler (default print)}

\item{interval}{timeframe of bars (availible "1min","2min","3min","5min","10min","15min","30min","hour","2hour","4hour","day","week","month")}

\item{depth}{depth of orderbook, number of bids and asks (1 <= depth <= 20)}
}
\description{
this function subscribes and unsubscribe for streaming
}
\details{
As described by the official Tinkoff Investments documentation.
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\seealso{
\code{\link{streamStatus}} \code{\link{streamClose}}  \code{\link{streamClient}}
}
\author{
Vyacheslav Arbuzov
}
