% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancelOrder.R
\name{cancelOrder}
\alias{cancelOrder}
\title{cancelOrder}
\usage{
cancelOrder(token = "", live = FALSE, orderId = "")
}
\arguments{
\item{token}{token from tinkoff}

\item{live}{live trading - TRUE or sandbox (paper) trading - FALSE}

\item{orderId}{id of order from getOrders}
}
\description{
cancelOrder
}
\details{
As described by the official Tinkoff Investments documentation
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
paper_token = 'your_sandbox_token_from_tcs_account'
cancelOrder(paper_token,'your_order_id')

live_token = 'your_live_token_from_tcs_account'
# remember that this command will close your order in live account !
cancelOrder(live_token,TRUE,'your_order_id')

}
\seealso{
\code{\link{getOrders}}
}
