% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancelOrder.R
\name{cancelOrder}
\alias{cancelOrder}
\title{tcsinvest}
\usage{
cancelOrder(token = "", orderId = "", sandbox = TRUE)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
