% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrencies.R
\name{getCurrencies}
\alias{getCurrencies}
\title{tcsinvest}
\usage{
getCurrencies(token = "", sandbox = TRUE)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
