% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getETFs.R
\name{getETFs}
\alias{getETFs}
\title{Download ETF data}
\usage{
getETFs(token = "", sandbox = TRUE)
}
\arguments{
\item{token}{token from Tinkoff account}

\item{sandbox}{paper (TRUE) or live (FALSE) trading}
}
\description{
Makes a request to the server, and returns an data.table object with information about available ETFs if successful.
}
\details{
As described by the official Tinkoff Investments documentation
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
token = 'your_token_from_tcs_account'
getETFs(token)
}
\seealso{
\code{\link{getCurrencies}} \code{\link{getBonds}} \code{\link{getStocks}}
}
\author{
Vyacheslav Arbuzov
}
