% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistoricalData.R
\name{getHistoricalData}
\alias{getHistoricalData}
\title{tcsinvest}
\usage{
getHistoricalData(
  token = "",
  figi = "",
  from = Sys.Date() - 2,
  to = Sys.Date(),
  interval = "hour",
  sandbox = TRUE
)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
