% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOperations.R
\name{getOperations}
\alias{getOperations}
\title{tcsinvest}
\usage{
getOperations(
  token = "",
  from = Sys.Date() - 2,
  to = Sys.Date(),
  sandbox = TRUE
)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
