% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrderBook.R
\name{getOrderBook}
\alias{getOrderBook}
\title{tcsinvest}
\usage{
getOrderBook(token = "", figi = "", depth = NULL, sandbox = TRUE)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
