% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbolInfo.R
\name{getSymbolInfo}
\alias{getSymbolInfo}
\title{tcsinvest}
\usage{
getSymbolInfo(token = "", figi = "", sandbox = TRUE)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
