% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTickerInfo.R
\name{getTickerInfo}
\alias{getTickerInfo}
\title{tcsinvest}
\usage{
getTickerInfo(token = "", ticker = "", sandbox = TRUE)
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
