% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniverse.R
\name{getUniverse}
\alias{getUniverse}
\title{Download universe data (stocks,bonds,etfs,currencies)}
\usage{
getUniverse(token = "", live = FALSE)
}
\arguments{
\item{token}{token from Tinkoff account}

\item{live}{live trading - TRUE or sandbox (paper) trading - FALSE (default)}
}
\description{
Makes 4 request to the server, and returns an data.table object with information about available instruments in Tinkoff
}
\details{
this function for not from official list of functions.
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
live = FALSE
token = 'your_sandbox_token_from_tcs_account'
getUniverse(token,live)
}
\seealso{
\code{\link{getCurrencies}} \code{\link{getBonds}} \code{\link{getETFs}} \code{\link{getStocks}}
}
\author{
Vyacheslav Arbuzov
}
