% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limitOrder.R
\name{limitOrder}
\alias{limitOrder}
\title{Place limit order}
\usage{
limitOrder(
  token = "",
  figi = "",
  direction = NULL,
  lots = NULL,
  price = NULL,
  sandbox = TRUE
)
}
\arguments{
\item{token}{token from Tinkoff account}

\item{figi}{internal tinkoff code for instrument}

\item{direction}{"Buy" or "Sell"}

\item{lots}{number of lots to buy}

\item{price}{price of limit order}

\item{sandbox}{paper (TRUE) or live (FALSE) trading}
}
\description{
this function will place your order in to account.
}
\details{
As described by the official Tinkoff Investments documentation. If you want live trading, use sandbox=FALSE with live token
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
token = 'your_token_from_tcs_account'
# remember that this command will place limit order in your account !
limitOrder(token,figi='BBG005HLTYH9',direction='Buy',lots=1,price=100,sandbox = TRUE)
}
\seealso{
\code{\link{marketOrder}}
}
\author{
Vyacheslav Arbuzov
}
