% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandboxBalance.R
\name{sandboxBalance}
\alias{sandboxBalance}
\title{tcsinvest}
\usage{
sandboxBalance(token = "", balance = NULL, currency = "RUB")
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
