% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandboxDeleteAccount.R
\name{sandboxDeleteAccount}
\alias{sandboxDeleteAccount}
\title{Working with sandbox. Delete account}
\usage{
sandboxDeleteAccount(token = "")
}
\arguments{
\item{token}{token from Tinkoff account (only sandbox token)}
}
\description{
Delete sandbox account
}
\details{
If you need to create few money position, use this function for each position
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
token = 'your_token_from_tcs_account'
sandboxDeleteAccount(token)
}
\seealso{
\code{\link{sandboxRegister}}
}
\author{
Vyacheslav Arbuzov
}
