% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandboxDeletePositions.R
\name{sandboxDeletePositions}
\alias{sandboxDeletePositions}
\title{tcsinvest}
\usage{
sandboxDeletePositions(token = "")
}
\arguments{
\item{dataframe}{}
}
\value{
the valuet
}
\description{
function
}
\examples{
cancelOrder(token)
}
