% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandboxRegister.R
\name{sandboxRegister}
\alias{sandboxRegister}
\title{Working with sandbox. Register account}
\usage{
sandboxRegister(token = "")
}
\arguments{
\item{token}{token from Tinkoff account (only sandbox token)}
}
\description{
Register sandbox account
}
\details{
If you need to create few money position, use this function for each position
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
token = 'your_token_from_tcs_account'
sandboxRegister(token)
}
\seealso{
\code{\link{sandboxDeleteAccount}}
}
\author{
Vyacheslav Arbuzov
}
